+++ src/gui_label.cpp
@@ -27,7 +27,9 @@ GUILabel::GUILabel (const std::string& title, int x, int y, int width, int heigh
 void
 GUILabel::draw (GraphicContext* gc)
 {
-  gc->draw_string(m_x + 10, m_y + 15, m_title);
+  gc->draw_string(static_cast<float>(m_x) + 10,
+                  static_cast<float>(m_y) + 15,
+                  m_title);
 }
 
 /* EOF */
